ILoggerFactory adapter for Monolog log engine.
Tools to use Monolog as the underlying logging engine for a hierarchical
logger stack. Introduces the concept of logger inheritance to Monolog. This is
not true inheritance as implemented by log4php, but a copy on construct fake
that can be accomplished with a thin shim layer on top of Monolog.